Update and rename sesrdyhdfgh to PalindromeChecker
Write a Java program to check if a given string is a palindrome.Use a stack to help you compare characters from the beginning and end of the string. Do not use any String functionalities to reverse the String. Use Stack data structure. You can use (A string is called Palindrome if the reverse of the string is the same as the original string. Example: “racecar”. )